login vs 4.5
Now when user logs in and they go back to login page, it will redirect them to the account that they are on. so when they go back ot the search and then click back on the profile, it will direct them directly to their profile instead of back to login.